comparison of unsigned expression >= 0 is always true
fix by making file handle positive and undoing previous commit.

+++ emulator/src/flash_storage.h
@@ -72,7 +72,7 @@ class FlashStorage {
  private:
     bool isInitialized;
     bool isFileBacked;
-    uintptr_t fileHandle;
+    int fileHandle;
     uintptr_t mappingHandle;
 
     bool mapFile(const char *filename);
